changed type of 'n' to std::size_t

+++ strings/rabin_karp.cpp
@@ -22,7 +22,7 @@ namespace string_search {
  * \param[in] n length of substring to hash
  * \returns hash integer
  */
-int64_t create_hash(const std::string& s, int n) {
+int64_t create_hash(const std::string& s, size_t n) {
     int64_t result = 0;
     for (int i = 0; i < n; ++i) {
         result += (int64_t)(s[i] * (int64_t)pow(PRIME, i));
